Fixed merge and added Navbar as its own component
