+++ src/components/Badge/Badge.jsx
@@ -22,7 +22,7 @@ const Badge = (props) => {
 
 
   return (
-    <Card style={{ backgroundColor: '#fafafa', borderRadius: 0 }}>
+    <Card style={{ backgroundColor: '#fafafa', borderRadius: 0 }} id="badgesearned">
       <CardHeader tag="h3">
         Badges
       </CardHeader>
