+++ src/components/Dashboard/board.css
@@ -122,6 +122,7 @@
   position: absolute;
   top: 0px;
   left: 0px;
+  cursor: pointer;
 }
 
 .badge-list {
